Fix duplicate option entries when including a YAML file
Issue XFITTER-87

This was caused by some nonintuitive behavior of yaml-cpp,
which led to duplicate keys in YAML maps.

Now locally-defined options override options included from a file,
issuing a warning.

However, because it is in general difficult to check if two
YAML nodes are equal, we currently only handle cases when the map key
is a scalar.