Fix evolutions not getting default decomposition
If no "decomposition" is provided to an evolution, it should fall back
to the default decomposition, as given by "DefaultDecomposition". If
there is no "DefaultDecomposition", it will try to get a decomposition
with the name "default".
It used to give an error if a "decomposition" is absent, which is
incorrect.