Use internal grid for LHAPDF6 output if possible
Internal grid will be used if evolution getXgrid() and getQgrid() return
non-empty arrays. However, this is not yet implemented for any
evolution.

If any Q- or X- grid parameters are provided under WriteLHAPDF6,
internal grid of evolution will be ignored and a new grid will be
generated, according to the given parameters.

To override this behavior, option "preferInternalGrid" can be used.
Then, provided grid parameters are only used if the evolution cannot
provide its internal grid.